requirements/local.txt
# Local development dependencies go here
-r base.txt
django-sphinx==2.2.4
coverage==3.7.1
django-discover-runner==1.0
django-debug-toolbar==1.0.1
docutils==0.11
django-extensions==1.3.3
